# examples/vanilla-mfa/openapi/mfa-api.yaml
openapi: 3.1.0

info:
  title: Family Ties MFA (Micro-Frontend Adapter) API
  version: 1.0.0
  description: |
    REST API exposed by the Micro-Frontend Adapter (MFA) to the Angular MFE.
    The MFA forwards all requests transparently to the Family Ties microservice.
    No authentication, session management, or header transformation is performed
    in this vanilla configuration.
  contact:
    name: Arc-E-Tect
    url: https://arc-e-tect.com

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: persons
    description: Manage persons in the family tree (proxied to MS)
  - name: relationships
    description: Manage relationships between persons (proxied to MS)

paths:
  /v1/familyties:
    post:
      tags: [persons]
      operationId: addPerson
      summary: Add a new person to the family tree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/shared/PersonRequest.yaml'
      responses:
        '201':
          description: Person created successfully
          headers:
            Location:
              schema:
                type: string
              description: URL of the created person resource
          content:
            application/json:
              schema:
                $ref: './components/shared/PersonResponse.yaml'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'
        '502':
          description: Upstream microservice error
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'

  /v1/familyties/lastnames/{lastName}:
    get:
      tags: [persons]
      operationId: findByLastName
      summary: Find persons by last name
      parameters:
        - name: lastName
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9\\s'-]+$"
          example: Smith
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of persons with the given last name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './components/shared/PersonResponse.yaml'
        '404':
          description: No persons found
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'
        '502':
          description: Upstream microservice error
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'

    delete:
      tags: [persons]
      operationId: deletePerson
      summary: Delete a person by first name and last name
      parameters:
        - name: lastName
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9\\s'-]+$"
          example: Smith
        - name: firstname
          in: query
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9\\s'-]+$"
          example: John
      responses:
        '204':
          description: Person deleted successfully
        '404':
          description: Person not found
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'
        '502':
          description: Upstream microservice error
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'

  /v1/familyties/relationships:
    post:
      tags: [relationships]
      operationId: addRelationship
      summary: Add a relationship between two persons
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/shared/RelationshipRequest.yaml'
      responses:
        '201':
          description: Relationship created successfully
          content:
            application/json:
              schema:
                $ref: './components/shared/RelationshipResponse.yaml'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'
        '502':
          description: Upstream microservice error
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'

  /v1/familyties/relationships/{type}/lastnames/{lastName}:
    get:
      tags: [relationships]
      operationId: findRelations
      summary: Find all persons related to a last name by relationship type
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z]+$"
          example: parent
        - name: lastName
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9\\s'-]+$"
          example: Smith
      responses:
        '200':
          description: List of related persons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './components/shared/PersonResponse.yaml'
        '404':
          description: No relations found
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'
        '502':
          description: Upstream microservice error
          content:
            application/json:
              schema:
                $ref: './components/shared/ErrorResponse.yaml'

  /actuator/health:
    get:
      tags: [system]
      operationId: health
      summary: Health check endpoint
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: UP